Create RangeSelector and SwitchSelector custom components and use for filter modal selections. Refactor filter modals to use new custom components.
