+++ src/screens/TimelineScreen/PhotoFilter.tsx
@@ -1,14 +1,12 @@
-import { FontAwesome5 } from "@expo/vector-icons";
 import { useEffect } from "react";
-import { Pressable, StyleSheet } from "react-native";
 
-import { Text, View } from "../../components/Themed";
 import {
   setLogsWithPics,
   switchFilterByPic,
 } from "../../features/Filters/filter.slice";
 import { logSelectors } from "../../features/log/log.slice";
 import { useAppDispatch, useAppSelector } from "../../store";
+import { SwitchSelector } from "../FilterModal/SwitchSelector";
 
 export const PhotoFilter = () => {
   const dispatch = useAppDispatch();
@@ -37,20 +35,10 @@ export const PhotoFilter = () => {
   }, [filterByPic]);
 
   return (
-    <View>
-      <Pressable onPress={handlePress}>
-        <Text style={styles.categorySelector}>Photo</Text>
-      </Pressable>
-      <View style={{ marginLeft: 200, justifyContent: "flex-end" }}>
-        {filterByPic && <FontAwesome5 name="check" size={24} color="green" />}
-      </View>
-    </View>
+    <SwitchSelector
+      name="Photo"
+      handlePress={handlePress}
+      tickCondition={!filterByPic}
+    />
   );
 };
-
-const styles = StyleSheet.create({
-  categorySelector: {
-    fontSize: 20,
-    marginTop: 30,
-  },
-});
