+++ src/screens/TimelineScreen/ResetFilters.tsx
@@ -1,0 +1,12 @@
+import { SwitchSelector } from "../FilterModal/SwitchSelector";
+
+export const ResetFilters: React.FC<{
+  clearFilters: () => void;
+  activeFilter: boolean;
+}> = ({ clearFilters, activeFilter }) => (
+  <SwitchSelector
+    name="None"
+    handlePress={clearFilters}
+    tickCondition={activeFilter}
+  />
+);
