+++ src/screens/TimelineScreen/TimelineScreen.tsx
@@ -2,25 +2,25 @@ import { Ionicons } from "@expo/vector-icons";
 import { useState } from "react";
 import { Pressable, StyleSheet } from "react-native";
 import { Text } from "react-native-elements";
-import { SafeAreaView } from "react-native-safe-area-context";
 
 import { View } from "../../components/Themed";
-import { FilterModal } from "./FilterModal";
+import { FilterModal } from "../FilterModal/FilterModal";
 import { TimelineElement } from "./TimelineElement";
 
 export const TimelineScreen = () => {
   const [modalVisible, setModalVisible] = useState(false);
 
+  const FilterButton = () => (
+    <Pressable style={[styles.button]} onPress={() => setModalVisible(true)}>
+      <Ionicons name="filter" size={24} color="black" />
+      <Text style={styles.buttonText}>Filters</Text>
+    </Pressable>
+  );
+
   return (
     <View style={styles.container}>
       <View style={styles.filter}>
-        <Pressable
-          style={[styles.button]}
-          onPress={() => setModalVisible(true)}
-        >
-          <Ionicons name="filter" size={24} color="black" />
-          <Text style={styles.buttonText}>Filters</Text>
-        </Pressable>
+        <FilterButton />
         <FilterModal
           setModalVisible={setModalVisible}
           modalVisible={modalVisible}
